Moves MetadataInfo to its own file
